LFOPTEE-277 drivers: ele: Move header files to core/include/drivers
Move header file from drivers/ele/include folder to
core/include/drivers/ele/ folder for use by PTA
