+++ core/drivers/crypto/ele/utils_mem.c
@@ -2,10 +2,10 @@
 /*
  * Copyright 2023 NXP
  */
+#include <drivers/ele/utils_mem.h>
 #include <io.h>
 #include <mm/core_memprot.h>
 #include <string.h>
-#include <utils_mem.h>
 
 /*
  * Allocate cache aligned memory of given size in bytes.
