+++ core/include/drivers/ele/key_mgmt.h
