+++ core/include/drivers/ele/sign_verify.h
@@ -8,7 +8,7 @@
 #include <stddef.h>
 #include <stdint.h>
 #include <tee_api_types.h>
-#include <utils_mem.h>
+#include <drivers/ele/utils_mem.h>
 
 #define ELE_KEY_USAGE_SIGN_MSG	  0x00000400
 #define ELE_KEY_USAGE_VERIFY_MSG  0x00000800
