+++ core/include/drivers/ele/utils_mem.h
